big re-org moving things into new files. more logical.
