+++ benchmark_tools_test.py
@@ -2,11 +2,11 @@
 from string import ascii_letters
 import numpy as np
 import pandas as pd
-import benchmark_tools as bt
 import scipy.stats as ss
 from sklearn.metrics import brier_score_loss, log_loss, zero_one_loss
-from sklearn.preprocessing import OneHotEncoder, normalize
-from statsmodels.distributions.empirical_distribution import StepFunction
+import benchmark_tools as bt
+import classification as btc
+import util
 
 
 def test_t_EB(runs=100, trials=1000):
@@ -75,81 +75,20 @@ def test_get_mean_and_EB(runs=100, trials=1000):
     return pval_agg
 
 
-def test_one_hot():
-    n_labels = np.random.randint(low=1, high=10)
-    N = np.random.randint(low=0, high=10)
-
-    y = np.random.randint(low=0, high=n_labels, size=N)
-    z0 = bt.one_hot(y, n_labels)
-    assert(z0.dtype.kind == 'b' and z0.shape == (N, n_labels))
-
-    if N >= 1:
-        enc = OneHotEncoder(n_values=n_labels, sparse=False, dtype=bool)
-        z1 = enc.fit_transform(y[:, None])
-        assert(np.all(z0 == z1))
-
-
-def test_normalize():
-    n_labels = np.random.randint(low=1, high=10)
-    N = np.random.randint(low=0, high=10)
-
-    log_pred_prob = np.random.randn(N, n_labels)
-    z0 = bt.normalize(log_pred_prob)
-    assert(z0.shape == (N, n_labels))
-
-    if N >= 1:
-        z1 = np.log(normalize(np.exp(log_pred_prob), norm='l1', axis=1))
-        assert(np.allclose(z0, z1))
-
-
-def epsilon_noise_test():
-    N = np.random.randint(low=0, high=10)
-    x = np.random.randn(N)
-
-    _, idx0 = np.unique(x, return_inverse=True)
-    _, idx1 = np.unique(bt.epsilon_noise(x), return_inverse=True)
-    assert(np.all(idx0 == idx1))
-
-
-def eval_step_func_test():
-    N = np.random.randint(low=0, high=10)
-
-    xp = np.random.randn(N)
-    yp = np.random.randn(N)
-    ival = np.random.randn()
-
-    assume_sorted = np.random.rand() <= 0.5
-    if assume_sorted:
-        xp = np.sort(xp)
-
-    y_grid = bt.eval_step_func(xp, xp, yp, assume_sorted=assume_sorted)
-    assert(np.allclose(y_grid, yp))
-
-    N_test = np.random.randint(low=0, high=10)
-    x_grid = np.concatenate((xp, np.random.randn(N_test)))
-
-    y_grid = bt.eval_step_func(x_grid, xp, yp, ival)
-
-    SF = StepFunction(xp, yp, ival=ival, side='right', sorted=assume_sorted)
-    y_grid2 = SF(x_grid)
-
-    assert(np.allclose(y_grid, y_grid2))
-
-
 def hard_loss_binary_test():
     '''Also tests hard loss.'''
     n_labels = 2
     N = np.random.randint(low=1, high=10)
 
     y_bool = np.random.rand(N) <= 0.5
-    y_pred = bt.normalize(np.random.randn(N, n_labels))
-    loss = bt.hard_loss_binary(y_bool, y_pred)
+    y_pred = util.normalize(np.random.randn(N, n_labels))
+    loss = btc.hard_loss_binary(y_bool, y_pred)
 
-    act = bt.hard_loss_decision(y_pred, 1.0 - np.eye(n_labels))
+    act = btc.hard_loss_decision(y_pred, 1.0 - np.eye(n_labels))
     loss2 = zero_one_loss(y_bool.astype(int), act)
     assert(np.allclose(np.mean(loss), loss2))
 
-    loss2 = bt.hard_loss(y_bool, y_pred)
+    loss2 = btc.hard_loss(y_bool, y_pred)
     assert(np.allclose(loss, loss2))
 
 
@@ -158,20 +97,20 @@ def hard_loss_decision_test():
     n_act = np.random.randint(low=1, high=10)
     N = np.random.randint(low=0, high=10)
 
-    y_pred = bt.normalize(np.random.randn(N, n_labels))
+    y_pred = util.normalize(np.random.randn(N, n_labels))
 
-    act = bt.hard_loss_decision(y_pred, 1.0 - np.eye(n_labels))
+    act = btc.hard_loss_decision(y_pred, 1.0 - np.eye(n_labels))
     assert(np.all(np.argmax(y_pred, axis=1) == act))
 
     loss_mat = np.random.rand(n_labels, n_act)
-    act = bt.hard_loss_decision(y_pred, loss_mat)
+    act = btc.hard_loss_decision(y_pred, loss_mat)
 
     loss_mat = np.concatenate((loss_mat, np.ones((n_labels, 1))), axis=1)
-    act2 = bt.hard_loss_decision(y_pred, loss_mat)
+    act2 = btc.hard_loss_decision(y_pred, loss_mat)
     assert(np.all(act == act2))
 
     loss_mat = np.concatenate((loss_mat, np.zeros((n_labels, 1))), axis=1)
-    act2 = bt.hard_loss_decision(y_pred, loss_mat)
+    act2 = btc.hard_loss_decision(y_pred, loss_mat)
     assert(np.all(act2 == loss_mat.shape[1] - 1))
 
 
@@ -180,16 +119,16 @@ def log_loss_test():
     N = np.random.randint(low=1, high=10)
 
     y = np.random.randint(low=0, high=n_labels, size=N)
-    y_pred = bt.normalize(np.random.randn(N, n_labels))
+    y_pred = util.normalize(np.random.randn(N, n_labels))
 
     if n_labels >= 2:
-        loss = bt.log_loss(y, y_pred)
+        loss = btc.log_loss(y, y_pred)
         loss2 = log_loss(y, np.exp(y_pred), labels=xrange(n_labels))
         assert(np.allclose(np.mean(loss), loss2))
 
     with np.errstate(invalid='ignore', divide='ignore'):
-        pred = np.log(bt.one_hot(y, n_labels))
-    loss2 = bt.log_loss(y, pred)
+        pred = np.log(util.one_hot(y, n_labels))
+    loss2 = btc.log_loss(y, pred)
     assert(np.max(np.abs(loss2)) <= 1e-8)
 
 
@@ -198,17 +137,17 @@ def brier_loss_test():
     N = np.random.randint(low=1, high=10)
 
     y = np.random.randint(low=0, high=n_labels, size=N)
-    y_pred = bt.normalize(np.random.randn(N, n_labels))
+    y_pred = util.normalize(np.random.randn(N, n_labels))
 
-    loss = bt.brier_loss(y, y_pred, rescale=False)
+    loss = btc.brier_loss(y, y_pred, rescale=False)
     # sklearn learn is dumb and gets confused when only one class passed in
     if n_labels == 2 and np.std(y) >= 1e-8:
         loss2 = brier_score_loss(y == 1, np.exp(y_pred[:, 1]), pos_label=True)
         assert(np.allclose(np.mean(loss), 2.0 * loss2))
 
     with np.errstate(invalid='ignore', divide='ignore'):
-        pred = np.log(bt.one_hot(y, n_labels))
-    loss2 = bt.brier_loss(y, pred, rescale=False)
+        pred = np.log(util.one_hot(y, n_labels))
+    loss2 = btc.brier_loss(y, pred, rescale=False)
     assert(np.max(np.abs(loss2)) <= 1e-8)
 
 
@@ -217,9 +156,9 @@ def spherical_loss_test():
     N = np.random.randint(low=1, high=10)
 
     y = np.random.randint(low=0, high=n_labels, size=N)
-    y_pred = bt.normalize(np.random.randn(N, n_labels))
+    y_pred = util.normalize(np.random.randn(N, n_labels))
 
-    loss = bt.spherical_loss(y, y_pred, rescale=False)
+    loss = btc.spherical_loss(y, y_pred, rescale=False)
 
     # Check against the linear implementation
     pred_prob = np.exp(y_pred)
@@ -228,14 +167,14 @@ def spherical_loss_test():
     assert(np.allclose(loss, loss_0, equal_nan=True))
 
     with np.errstate(invalid='ignore', divide='ignore'):
-        pred = np.log(bt.one_hot(y, n_labels))
-    loss2 = bt.spherical_loss(y, pred, rescale=False)
+        pred = np.log(util.one_hot(y, n_labels))
+    loss2 = btc.spherical_loss(y, pred, rescale=False)
     assert(np.max(np.abs(loss2 + 1.0)) <= 1e-8)
 
     if n_labels >= 2:
         with np.errstate(invalid='ignore', divide='ignore'):
-            pred_prob = bt.normalize(np.log(1.0 - bt.one_hot(y, n_labels)))
-        loss2 = bt.spherical_loss(y, pred_prob, rescale=False)
+            pred_prob = util.normalize(np.log(1.0 - util.one_hot(y, n_labels)))
+        loss2 = btc.spherical_loss(y, pred_prob, rescale=False)
         assert(np.max(np.abs(loss2)) <= 1e-8)
 
 
@@ -249,23 +188,23 @@ def loss_summary_table_test():
 
     methods = np.random.choice(list(ascii_letters), n_methods, replace=False)
     ref = np.random.choice(methods)
-    metrics = bt.STD_MULTICLASS_LOSS
+    metrics = btc.STD_MULTICLASS_LOSS
     labels = xrange(n_labels)
 
     col_names = pd.MultiIndex.from_product([methods, labels],
-                                           names=[bt.METHOD, bt.LABEL])
+                                           names=[bt.METHOD, btc.LABEL])
     dat = np.random.randn(N, n_labels * len(methods))
     tbl = pd.DataFrame(data=dat, index=xrange(N), columns=col_names,
                        dtype=float)
 
     y = np.random.randint(low=0, high=n_labels, size=N)
-    loss_tbl = bt.loss_table(tbl, y, metrics_dict=metrics)
+    loss_tbl = btc.loss_table(tbl, y, metrics_dict=metrics)
     perf_tbl = bt.loss_summary_table(loss_tbl, ref, pairwise_CI=pairwise_CI,
                                      confidence=confidence)
     for metric, metric_f in metrics.iteritems():
-        loss_ref = metric_f(y, bt.normalize(tbl[ref].values))
+        loss_ref = metric_f(y, util.normalize(tbl[ref].values))
         for method in methods:
-            loss = metric_f(y, bt.normalize(tbl[method].values))
+            loss = metric_f(y, util.normalize(tbl[method].values))
             assert(np.allclose(loss_tbl[(metric, method)].values, loss,
                                equal_nan=True))
 
@@ -288,17 +227,14 @@ def loss_summary_table_test():
 np.random.seed(53634)
 
 for _ in xrange(1000):
-    test_one_hot()
-    test_normalize()
-    epsilon_noise_test()
     hard_loss_binary_test()
     hard_loss_decision_test()
     log_loss_test()
     brier_loss_test()
     spherical_loss_test()
     loss_summary_table_test()
-    eval_step_func_test()
-print 'Now running MC tests'
 
+print 'Now running MC tests'
 print test_t_EB()
-print test_get_mean_and_EB(trials=10000)
+print test_get_mean_and_EB()
+print 'passed'
