+++ boot_test.py
@@ -1,6 +1,6 @@
 # Ryan Turner (turnerry@iro.umontreal.ca)
 import numpy as np
-import benchmark_tools as bt
+from classification import curve_boot
 import perf_curves as pc
 
 np.random.seed(3563)
@@ -21,34 +21,35 @@ for rr in xrange(runs):
 
     x_curve, y_curve, _ = pc.roc_curve(y_true[N:], y_score[N:, 1])
     ref = pc.auc_trapz(x_curve, y_curve)
-    summary, _ = bt.curve_boot(y_true[:N], y_score[:N, :],
-                               default_summary_ref=ref,
-                               curve_f=pc.roc_curve, summary_f=pc.auc_trapz,
-                               n_boot=1000, confidence=confidence)
+    summary, _ = curve_boot(y_true[:N], y_score[:N, :],
+                            default_summary_ref=ref,
+                            curve_f=pc.roc_curve, summary_f=pc.auc_trapz,
+                            n_boot=1000, confidence=confidence)
     auc[rr, 0], EB, pval[rr, 0] = summary
     valid[rr, 0] = np.abs(ref - auc[rr, 0]) <= EB
 
     x_curve, y_curve, _ = pc.recall_precision_curve(y_true[N:], y_score[N:, 1])
     ref = pc.auc_left(x_curve, y_curve)
-    summary, _ = bt.curve_boot(y_true[:N], y_score[:N, :],
-                               default_summary_ref=ref,
-                               curve_f=pc.recall_precision_curve,
-                               summary_f=pc.auc_left,
-                               n_boot=1000, confidence=confidence)
+    summary, _ = curve_boot(y_true[:N], y_score[:N, :],
+                            default_summary_ref=ref,
+                            curve_f=pc.recall_precision_curve,
+                            summary_f=pc.auc_left,
+                            n_boot=1000, confidence=confidence)
     auc[rr, 1], EB, pval[rr, 1] = summary
     valid[rr, 1] = np.abs(ref - auc[rr, 1]) <= EB
 
     x_curve, y_curve, _ = pc.prg_curve(y_true[N:], y_score[N:, 1])
     ref = pc.auc_left(x_curve, y_curve)
-    summary, _ = bt.curve_boot(y_true[:N], y_score[:N, :],
-                               default_summary_ref=ref,
-                               curve_f=pc.prg_curve, summary_f=pc.auc_left,
-                               n_boot=1000, confidence=confidence)
+    summary, _ = curve_boot(y_true[:N], y_score[:N, :],
+                            default_summary_ref=ref,
+                            curve_f=pc.prg_curve, summary_f=pc.auc_left,
+                            n_boot=1000, confidence=confidence)
     auc[rr, 2], EB, pval[rr, 2] = summary
     valid[rr, 2] = np.abs(ref - auc[rr, 2]) <= EB
-print np.mean(valid, axis=0)
-print np.mean(auc, axis=0)
-print np.mean(pval <= 0.05, axis=0)
+print 'againt N-big'
+print 'P EB valid', np.mean(valid, axis=0)
+print 'mean AUC', np.mean(auc, axis=0)
+print 'P p-val sig', np.mean(pval <= 0.05, axis=0)
 
 auc = np.zeros((runs, 3))
 valid = np.zeros((runs, 3))
@@ -57,24 +58,25 @@ for rr in xrange(runs):
     y_true = np.random.rand(N) <= p
     y_score = np.random.randn(N, 2)
 
-    summary, _ = bt.curve_boot(y_true, y_score, default_summary_ref=0.5,
-                               curve_f=pc.roc_curve, summary_f=pc.auc_trapz,
-                               n_boot=1000, confidence=confidence)
+    summary, _ = curve_boot(y_true, y_score, default_summary_ref=0.5,
+                            curve_f=pc.roc_curve, summary_f=pc.auc_trapz,
+                            n_boot=1000, confidence=confidence)
     auc[rr, 0], EB, pval[rr, 0] = summary
     valid[rr, 0] = np.abs(0.5 - auc[rr, 0]) <= EB
 
-    summary, _ = bt.curve_boot(y_true, y_score, default_summary_ref=p,
-                               curve_f=pc.recall_precision_curve,
-                               summary_f=pc.auc_left,
-                               n_boot=1000, confidence=confidence)
+    summary, _ = curve_boot(y_true, y_score, default_summary_ref=p,
+                            curve_f=pc.recall_precision_curve,
+                            summary_f=pc.auc_left,
+                            n_boot=1000, confidence=confidence)
     auc[rr, 1], EB, pval[rr, 1] = summary
     valid[rr, 1] = np.abs(p - auc[rr, 1]) <= EB
 
-    summary, _ = bt.curve_boot(y_true, y_score, default_summary_ref=0.0,
-                               curve_f=pc.prg_curve, summary_f=pc.auc_left,
-                               n_boot=1000, confidence=confidence)
+    summary, _ = curve_boot(y_true, y_score, default_summary_ref=0.0,
+                            curve_f=pc.prg_curve, summary_f=pc.auc_left,
+                            n_boot=1000, confidence=confidence)
     auc[rr, 2], EB, pval[rr, 2] = summary
     valid[rr, 2] = np.abs(auc[rr, 2]) <= EB
-print np.mean(valid, axis=0)
-print np.mean(auc, axis=0)
-print np.mean(pval <= 0.05, axis=0)
+print 'just noise'
+print 'P EB valid', np.mean(valid, axis=0)
+print 'mean AUC', np.mean(auc, axis=0)
+print 'P p-val sig', np.mean(pval <= 0.05, axis=0)
