+++ util_test.py
@@ -1,0 +1,75 @@
+# Ryan Turner (turnerry@iro.umontreal.ca)
+import numpy as np
+from sklearn.preprocessing import OneHotEncoder, normalize
+from statsmodels.distributions.empirical_distribution import StepFunction
+import util
+
+
+def test_one_hot():
+    n_labels = np.random.randint(low=1, high=10)
+    N = np.random.randint(low=0, high=10)
+
+    y = np.random.randint(low=0, high=n_labels, size=N)
+    z0 = util.one_hot(y, n_labels)
+    assert(z0.dtype.kind == 'b' and z0.shape == (N, n_labels))
+
+    if N >= 1:
+        enc = OneHotEncoder(n_values=n_labels, sparse=False, dtype=bool)
+        z1 = enc.fit_transform(y[:, None])
+        assert(np.all(z0 == z1))
+
+
+def test_normalize():
+    n_labels = np.random.randint(low=1, high=10)
+    N = np.random.randint(low=0, high=10)
+
+    log_pred_prob = np.random.randn(N, n_labels)
+    z0 = util.normalize(log_pred_prob)
+    assert(z0.shape == (N, n_labels))
+
+    if N >= 1:
+        z1 = np.log(normalize(np.exp(log_pred_prob), norm='l1', axis=1))
+        assert(np.allclose(z0, z1))
+
+
+def epsilon_noise_test():
+    N = np.random.randint(low=0, high=10)
+    x = np.random.randn(N)
+
+    _, idx0 = np.unique(x, return_inverse=True)
+    _, idx1 = np.unique(util.epsilon_noise(x), return_inverse=True)
+    assert(np.all(idx0 == idx1))
+
+
+def eval_step_func_test():
+    N = np.random.randint(low=0, high=10)
+
+    xp = np.random.randn(N)
+    yp = np.random.randn(N)
+    ival = np.random.randn()
+
+    assume_sorted = np.random.rand() <= 0.5
+    if assume_sorted:
+        xp = np.sort(xp)
+
+    y_grid = util.eval_step_func(xp, xp, yp, assume_sorted=assume_sorted)
+    assert(np.allclose(y_grid, yp))
+
+    N_test = np.random.randint(low=0, high=10)
+    x_grid = np.concatenate((xp, np.random.randn(N_test)))
+
+    y_grid = util.eval_step_func(x_grid, xp, yp, ival)
+
+    SF = StepFunction(xp, yp, ival=ival, side='right', sorted=assume_sorted)
+    y_grid2 = SF(x_grid)
+
+    assert(np.allclose(y_grid, y_grid2))
+
+np.random.seed(75675)
+
+for _ in xrange(1000):
+    test_one_hot()
+    test_normalize()
+    epsilon_noise_test()
+    eval_step_func_test()
+print 'passed'
